+++ modules/edit-inline.module.ts
@@ -1,10 +1,10 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { FormsModule } from '@angular/forms';
 import { AutofocusDirective } from './autofocus.directive';
 import { NgxEditInlineComponent } from './edit-inline.component';
+import { ReactiveFormsModule, FormsModule } from '@angular/forms';
 @NgModule({
-  imports:      [ CommonModule, FormsModule],
+  imports:      [CommonModule, ReactiveFormsModule, FormsModule],
   declarations: [ NgxEditInlineComponent, AutofocusDirective ],
   exports: [ NgxEditInlineComponent, AutofocusDirective ]
 })